+++ lib/hunspell_spell_check.dart
@@ -2,6 +2,11 @@ export 'src/custom_dictionary.dart'
     show CustomDictionary, CustomDictionaryStore, FileCustomDictionaryStore;
 export 'src/flutter_spell_check.dart';
 export 'src/hunspell_service.dart'
-    show AbstractAssetLoader, AssetLoader, HunspellService;
+    show
+        AbstractAssetLoader,
+        AssetLoader,
+        HunspellInitException,
+        HunspellService,
+        hunspellPathFor;
 export 'src/spell_check_options.dart';
 export 'src/spell_checker.dart';